+++ NEWS.md
@@ -10,6 +10,8 @@ Version 1.0.2b0 - dev
 - NEW: copy and transform support for `PDFUNDERLAY`, `DWFUNDERLAY` and `DGNUNDERLAY`
 - CHANGE: default flags for `UNDERLAY` entities is now 10 (underlay is on, adjust for background)
 - BUGFIX: fix ownership of sub-entities of `INSERT` and `POLYLINE` entities
+- BUGFIX: [#830](https://github.com/mozman/ezdxf/issues/830)
+  estimation of MTEXT column width when only white-spaces are present
 
 Version 1.0.1 - 2023-01-14
 --------------------------