+++ NEWS.md
@@ -8,6 +8,8 @@ Version 1.0.2b0 - dev
 - Release notes: https://ezdxf.mozman.at/release-v1-0.html
 - NEW: `Drawing.validate()` also prints report of resolved issues
 - NEW: copy and transform support for `PDFUNDERLAY`, `DWFUNDERLAY` and `DGNUNDERLAY`
+- NEW: [#832](https://github.com/mozman/ezdxf/issues/832) 
+  support for elliptic arcs in proxy graphics 
 - CHANGE: default flags for `UNDERLAY` entities is now 10 (underlay is on, adjust for background)
 - BUGFIX: fix ownership of sub-entities of `INSERT` and `POLYLINE` entities
 - BUGFIX: [#830](https://github.com/mozman/ezdxf/issues/830)